LPS-196245 Delete DDMDataProviderConfigurationActivator and move configuration to its usage
