--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/data/provider/configuration/activator/DDMDataProviderConfigurationActivator.java
@@ -1,39 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.dynamic.data.mapping.internal.data.provider.configuration.activator;
-
-import com.liferay.dynamic.data.mapping.data.provider.configuration.DDMDataProviderConfiguration;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-
-import java.util.Map;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Modified;
-
-/**
- * @author Leonardo Barros
- */
-@Component(
-	configurationPid = "com.liferay.dynamic.data.mapping.data.provider.configuration.DDMDataProviderConfiguration",
-	service = DDMDataProviderConfigurationActivator.class
-)
-public class DDMDataProviderConfigurationActivator {
-
-	public DDMDataProviderConfiguration getDDMDataProviderConfiguration() {
-		return _ddmDataProviderConfiguration;
-	}
-
-	@Activate
-	@Modified
-	protected void activate(Map<String, Object> properties) {
-		_ddmDataProviderConfiguration = ConfigurableUtil.createConfigurable(
-			DDMDataProviderConfiguration.class, properties);
-	}
-
-	private volatile DDMDataProviderConfiguration _ddmDataProviderConfiguration;
-
-}
